feat(core)!: SWG-CORE-01 green — the three canonical index fields are u64
Closes H3. `MasterBar.index`, `ImportWarning::TrackNameInvalidUtf8.track_index`,
and `ImportWarning::TempoApproximated.bar_index` go from `usize` to `u64`, so a
score written on a 64-bit host means the same thing on a 32-bit one. Spec §1.2
forbids platform-sized integers in hashed or serialized state and `Score`
derives `Hash`; until now the portable exact format was not portable.

**u64, not u32**, and the reason is semantic rather than defensive. The fields
are public, values above `u32::MAX` are already inhabited on a 64-bit host, and
`MasterBar.index` is a stored exact fact that need not agree with its vector
position (H4). `u32` would not merely have removed a platform dependence — it
would have shrunk a canonical domain that already has values in it. Step A's
witnesses are what make that a check rather than a claim.

## `index_from_ordinal`

One new public function in `core::score`, and the only sanctioned way to turn a
vector position into a canonical index. Importers legitimately derive the
second from the first; the point is that every such crossing is now named and
greppable rather than implicit. It is total and lossless — `usize` is at most
64 bits on every target Rust supports — and there is deliberately no inverse,
because the inverse is not total and must not be written by accident. Twenty
call sites, all of them constructors that were already numbering bars from an
`enumerate()`.

## Three types the migration had to pull along

Not scope creep — the alternative at each is a truncating `u64` → `usize`
conversion on a 32-bit target, which is the defect being fixed:

- `dump::NormBar.index` copies the stored index into a `Serialize` type;
- `SemanticPathSegment::MasterBar.index` and `::Bar.index` annotate diff paths
  with the stored index. The `ordinal` beside each is an operational position
  and stays `usize` — the two sit in the same struct and now have visibly
  different widths, which is the correct shape;
- `ui_core::analysis::Section.bar_start` / `.bar_end` are copied straight from
  `MasterBar.index`; `bar_count()` follows them to `u64`.

## What did not change

No operational `usize` was touched: `Vec` positions, lengths, and slice
indices are the same as before, including MIDI's `MAX_MASTER_BARS` resource
bound, which is an operational limit and not an argument about canonical
width. `nearest_micros` stays `u32` and step B has a test saying so. No index
became derived from a position, no warning payload became a lookup, no range
invariant was added, and nothing truncates, saturates, or clamps.

`cli::bar_at_tick` gains a comment rather than a fix: its two branches return
a stored index and a last ordinal respectively, a pre-existing conflation that
this task only makes same-width, not correct. It is display-only.

## Steps A and B

Step A's characterization is green, and the only edits to it are fixture
parameter types — every assertion is byte-identical to the pre-migration
commit. Its `#[cfg(target_pointer_width = "64")]` gates are gone, because the
migration is exactly what removed their reason to exist: those four assertions
now run on every target rather than on the one where `usize` happened to be
wide enough. Step B's contract compiles and passes, including the two
`>u32::MAX` cases that never had a `cfg`.

Verified: 1339 tests green across core, swang, pattern, cli, and ui-core;
clippy `-D warnings` and `cargo fmt --check` clean workspace-wide, cockpit,
preview, and plugin included (type-checked with `--all-targets` against a stub
pkg-config entry, since libasound is absent here). The two `fuzz/` targets that
build a `MasterBar` are edited identically to sites that do compile, but
`fuzz/` pins nightly and could not be built locally — CI covers it.

Refs SWG-CORE-01

+++ cockpit/src/lib.rs
@@ -4044,7 +4044,7 @@ mod tests {
         const BAR: u32 = 3840;
         let master_bars = (0..bars)
             .map(|i| MasterBar {
-                index: i as usize,
+                index: u64::from(i),
                 tick_range: TickRange::new(Ticks(i * BAR), Ticks((i + 1) * BAR)).expect("range"),
                 time_signature: TimeSignature::new(4, 4).expect("4/4"),
                 tempo: Tempo::from_bpm_integer(120).expect("120"),
